_scripts/requirements.txt
moz.l10n ~= 0.2.1
tomli-w ~= 1.0
